msm: peripheral-loader: Log clk-ready IRQ to help with debugging
This log message denotes the receipt of "clock ready" interrupt
which is received after a subsystem is brought up.
This interrupt is an indication to the kernel that the subsystem
is capable of voting for it's own resources, and that the kernel can
remove it's proxy votes for these resources.

Change-Id: Ib8a89e9e1ec2a547af4bd969ee31c94c42883d57
Signed-off-by: Deepak Katragadda <dkatraga@codeaurora.org>
Signed-off-by: Dan Trifan <jimsth@gmail.com>