msm: clock-generic: Add a combined mux_div clock
This new clock type is intended to handle certain special cases which
cannot be implemented with the existing mux and divider clock types.

Hardware may support switching both the parent clock and the divider
at the same time. This avoids generating intermediate frequencies from
either the old parent clock and new divider or new parent clock and old
divider combinations.

Extend the "safe parent" concept of the mux clock to a "safe freq".
"safe freq" is guaranteed to be within the fmax requirements for the
lowest voltage level, so it is safe to switch to at any time. It must be
defined when one of the parent clocks must be disabled in order to switch
rates.

Change-Id: I0163be99f2d14b38ce2c2d885df4d68e5c564dde
Signed-off-by: Patrick Daly <pdaly@codeaurora.org>
Signed-off-by: Dan Trifan <jimsth@gmail.com>