msm: clock-krait-8974: Run force reinit sequence for all CPUs
The HFPLL and krait mux force reinit sequence needs to be run for all CPUs
irrespective of whether they are online or offline. This is necessary to
ensure that when offline CPUs are brought online later on, they are brought
up at a known frequency.

To be able to do the above, we also have to remove the prepare count check
in krait core clock set rate ops. It's safe to do so since that check
wasn't really providing any additional protection on top of the checks
already done in the cpufreq driver.

Change-Id: Ia73459ee03d339c948d7f2ea9516df1c0d9b8bd0
Signed-off-by: Saravana Kannan <skannan@codeaurora.org>
Signed-off-by: Dan Trifan <jimsth@gmail.com>