arm/dt: msm-iommu-v1: don't gate XO shutdown
The MMSSNOC AXI clock is controlled by the apps clock driver,
therefore there is no concept of "active-only". If the clock
is on, it will stay on irrespective of whether the apps
processor is running or power-collapsed. Drivers need to
ensure that their requests are removed before entering a
low-power mode that is expected to lead to XO shutdown.
Therefore there is no point in including an active-only flag,
and drivers should not rely on this flag taking care of
turning off this clock, because that can't happen.

CRs-fixed: 521149
Change-Id: I2bb3eb1e223d13ef097619e64d64030482c026fe
Signed-off-by: Susheel Khiani <skhiani@codeaurora.org>
Signed-off-by: Dan Trifan <jimsth@gmail.com>