ASoC: msm: qdsp6v2: Update the topology if there is no ACDB
The topology for ASM and ADM is passed down from ACDB to adsp.
If there is no ACDB, then the topology ID is zero. In the current
implementation, whenever the topolpgy ID is zero, then the default
topology is used for asm and adm. The problem arises for msm8x10
because it support only null/none topolgy. It causes audio playback
failure in factory test mode when ACDB loader is disabled by factory
test application. Update the topology to null/none from default to
cover all the platforms.

CRs-fixed: 538213

Change-Id: I8cc694d731284294fb2f8341b370e0dfeb2a7d9b
Signed-off-by: Kuirong Wang <kuirongw@codeaurora.org>
Signed-off-by: Dan Trifan <jimsth@gmail.com>