Revert "msm: tty: update receive room just before writing data to the ldisc"
This reverts commit 913565a83d74fe1da0914e8c355345bc616131c4.

This change seems to be a hack introduced for some corner cases seen on
certain old fusion targets. The change is no longer applicable and is
being reverted to bring the driver closer to the upstream implementation.
The presence of this causes several merge conflicts when pulling upstream
fixes. It has already been reverted on mainline.

Change-Id: I9c08c96d77f478d21ca0b91ef2a8a3d5bf06cb00
Signed-off-by: Syed Rameez Mustafa <rameezmustafa@codeaurora.org>
Signed-off-by: Dan Trifan <jimsth@gmail.com>