power: max17048: power down if the battery voltage is too low
The battery could be not charged depends on system load even
if the charger is plugged. Since application doesn't power
down the device when the charger is plugged even if battery
level reach to 0%. As a result, the battery could be severely
drained under this condition.
To avoid sudden power loss by UVLO, monitor the battery voltage
and power down the device if the battery voltage drop below the
threshold.

Change-Id: I88dd068bdfb5aa39c8c35511432bf5a10bc70db5
Signed-off-by: choongryeol.lee <choongryeol.lee@lge.com>
Signed-off-by: Dan Trifan <jimsth@gmail.com>