ASoC: wcd9xxx: Cancel headset detect work while in IRQ handler
When the headset is inserted and removed with in no time, removal
interrupt handling is missed.

The reason behind missing the removal interrupt is for the insertion
interrupt a work queue is scheduled to detect the event type
there is latency in scheduling the work queue as a result the
removal IRQ followed is missed to process but the UI is not
notified about the IRQ miss and UI still thinks the headset is
connected.

Cancel the hs detect work queue as soon as entered the IRQ
and do not wait to cancel the work queue until the insertion or
removal event is detected

CRs-Fixed: 552760
Change-Id: I5c5fd8c55cca6a7137c389c19843c5178c8c9be8
Signed-off-by: Santosh Mardi <gsantosh@codeaurora.org>
Signed-off-by: Dan Trifan <jimsth@gmail.com>