nf: IDLETIMER: time-stamp and suspend/resume handling.
Message notifications contains an additional timestamp field in nano seconds.
The expiry time for the timers are modified during suspend/resume.
If timer was supposed to expire while the system is suspended then a
notification is sent when it resumes with the timestamp of the scheduled expiry.

Removes the race condition for multiple work scheduled.

Signed-off-by: Ruchi Kandoi <kandoiruchi@google.com>
Signed-off-by: Dan Trifan <jimsth@gmail.com>